Calculate CO2 GPM for PHEV uncombined
The DoE data set field for co2GpmFuel1/2 reflects a “utility factor”
for PHEV. Therefore, it isn’t the entire CO2 GPM for fuel type 1 or 2
separately; it is reduced by the portion of electricity usage. The
change in this commit separates the calculation to be purely for each
fuel type instead of combining electric and the primary fuel type.